+++ src/i18n/SerialTest_zh_CN.ts
@@ -10,7 +10,7 @@
     </message>
     <message>
         <location filename="../ui/controlitem.ui" line="63"/>
-        <location filename="../controlitem.cpp" line="238"/>
+        <location filename="../controlitem.cpp" line="235"/>
         <source>Command</source>
         <translation>命令</translation>
     </message>
@@ -74,7 +74,7 @@
     <message>
         <location filename="../ui/controlitem.ui" line="280"/>
         <location filename="../ui/controlitem.ui" line="311"/>
-        <location filename="../controlitem.cpp" line="138"/>
+        <location filename="../controlitem.cpp" line="136"/>
         <source>Prefix</source>
         <translation>前缀</translation>
     </message>
@@ -99,7 +99,7 @@
     <message>
         <location filename="../ui/controlitem.ui" line="328"/>
         <location filename="../ui/controlitem.ui" line="359"/>
-        <location filename="../controlitem.cpp" line="146"/>
+        <location filename="../controlitem.cpp" line="144"/>
         <source>Suffix</source>
         <translation>后缀</translation>
     </message>
@@ -148,72 +148,72 @@
     </message>
     <message>
         <location filename="../ui/ctrltab.ui" line="64"/>
-        <location filename="../ctrltab.cpp" line="87"/>
-        <location filename="../ctrltab.cpp" line="113"/>
+        <location filename="../ctrltab.cpp" line="81"/>
+        <location filename="../ctrltab.cpp" line="107"/>
         <source>Import</source>
         <translation>导入</translation>
     </message>
     <message>
         <location filename="../ui/ctrltab.ui" line="71"/>
-        <location filename="../ctrltab.cpp" line="146"/>
-        <location filename="../ctrltab.cpp" line="167"/>
+        <location filename="../ctrltab.cpp" line="139"/>
+        <location filename="../ctrltab.cpp" line="166"/>
         <source>Export</source>
         <translation>导出</translation>
     </message>
     <message>
-        <location filename="../ctrltab.cpp" line="89"/>
+        <location filename="../ctrltab.cpp" line="84"/>
         <source>Paste the exported data in the box.</source>
         <translation>将导出的数据粘贴到此框中。</translation>
     </message>
     <message>
-        <location filename="../ctrltab.cpp" line="93"/>
-        <location filename="../ctrltab.cpp" line="160"/>
+        <location filename="../ctrltab.cpp" line="88"/>
+        <location filename="../ctrltab.cpp" line="159"/>
         <source>Done</source>
         <translation>完成</translation>
     </message>
     <message>
-        <location filename="../ctrltab.cpp" line="120"/>
+        <location filename="../ctrltab.cpp" line="113"/>
         <source>Import Control Panel</source>
         <translation>导入控制面板</translation>
     </message>
     <message>
-        <location filename="../ctrltab.cpp" line="138"/>
-        <location filename="../ctrltab.cpp" line="150"/>
-        <location filename="../ctrltab.cpp" line="172"/>
-        <location filename="../ctrltab.cpp" line="186"/>
+        <location filename="../ctrltab.cpp" line="131"/>
+        <location filename="../ctrltab.cpp" line="143"/>
+        <location filename="../ctrltab.cpp" line="171"/>
+        <location filename="../ctrltab.cpp" line="189"/>
         <source>Info</source>
         <translation>信息</translation>
     </message>
     <message>
-        <location filename="../ctrltab.cpp" line="138"/>
-        <location filename="../ctrltab.cpp" line="186"/>
+        <location filename="../ctrltab.cpp" line="131"/>
+        <location filename="../ctrltab.cpp" line="189"/>
         <source>Successed!</source>
         <translation>成功！</translation>
     </message>
     <message>
-        <location filename="../ctrltab.cpp" line="138"/>
-        <location filename="../ctrltab.cpp" line="186"/>
+        <location filename="../ctrltab.cpp" line="131"/>
+        <location filename="../ctrltab.cpp" line="189"/>
         <source>Failed!</source>
         <translation>失败！</translation>
     </message>
     <message>
-        <location filename="../ctrltab.cpp" line="150"/>
-        <location filename="../ctrltab.cpp" line="172"/>
+        <location filename="../ctrltab.cpp" line="143"/>
+        <location filename="../ctrltab.cpp" line="171"/>
         <source>Please add item first</source>
         <translation>请先添加控制项</translation>
     </message>
     <message>
-        <location filename="../ctrltab.cpp" line="154"/>
+        <location filename="../ctrltab.cpp" line="147"/>
         <source>Copy all text in this box and save it to somewhere.</source>
         <translation>将此文本框中的内容复制并保存。</translation>
     </message>
     <message>
-        <location filename="../ctrltab.cpp" line="155"/>
+        <location filename="../ctrltab.cpp" line="148"/>
         <source>To import, click the Import button, then paste the text back.</source>
         <translation>需要导入时，单击“导入”按钮，将复制的内容粘贴回来即可。</translation>
     </message>
     <message>
-        <location filename="../ctrltab.cpp" line="178"/>
+        <location filename="../ctrltab.cpp" line="177"/>
         <source>Export Control Panel</source>
         <translation>导出控制面板</translation>
     </message>
@@ -278,7 +278,7 @@
     <message>
         <location filename="../ui/datatab.ui" line="266"/>
         <location filename="../ui/datatab.ui" line="320"/>
-        <location filename="../datatab.cpp" line="129"/>
+        <location filename="../datatab.cpp" line="175"/>
         <source>Suffix</source>
         <translation>后缀</translation>
     </message>
@@ -349,46 +349,46 @@
         <translation>设置</translation>
     </message>
     <message>
-        <location filename="../datatab.cpp" line="78"/>
-        <location filename="../datatab.cpp" line="219"/>
-        <location filename="../datatab.cpp" line="242"/>
+        <location filename="../datatab.cpp" line="122"/>
+        <location filename="../datatab.cpp" line="265"/>
+        <location filename="../datatab.cpp" line="288"/>
         <source>Info</source>
         <translation>信息</translation>
     </message>
     <message>
-        <location filename="../datatab.cpp" line="78"/>
+        <location filename="../datatab.cpp" line="122"/>
         <source>is not a valid encoding.</source>
         <translation>不是有效编码名。</translation>
     </message>
     <message>
-        <location filename="../datatab.cpp" line="160"/>
+        <location filename="../datatab.cpp" line="206"/>
         <source>Rx</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../datatab.cpp" line="167"/>
+        <location filename="../datatab.cpp" line="213"/>
         <source>Tx</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../datatab.cpp" line="203"/>
+        <location filename="../datatab.cpp" line="249"/>
         <source>Export received data</source>
         <translation>导出已接收数据</translation>
     </message>
     <message>
-        <location filename="../datatab.cpp" line="219"/>
-        <location filename="../datatab.cpp" line="242"/>
+        <location filename="../datatab.cpp" line="265"/>
+        <location filename="../datatab.cpp" line="288"/>
         <source>Successed!</source>
         <translation>成功！</translation>
     </message>
     <message>
-        <location filename="../datatab.cpp" line="219"/>
-        <location filename="../datatab.cpp" line="242"/>
+        <location filename="../datatab.cpp" line="265"/>
+        <location filename="../datatab.cpp" line="288"/>
         <source>Failed!</source>
         <translation>失败！</translation>
     </message>
     <message>
-        <location filename="../datatab.cpp" line="226"/>
+        <location filename="../datatab.cpp" line="272"/>
         <source>Export sended data</source>
         <translation>导出已发送数据</translation>
     </message>
@@ -397,7 +397,7 @@
     <name>DeviceTab</name>
     <message>
         <location filename="../ui/devicetab.ui" line="34"/>
-        <location filename="../devicetab.cpp" line="265"/>
+        <location filename="../devicetab.cpp" line="268"/>
         <source>Refresh</source>
         <translation>刷新</translation>
     </message>
@@ -548,7 +548,7 @@
     </message>
     <message>
         <location filename="../ui/devicetab.ui" line="268"/>
-        <location filename="../devicetab.cpp" line="163"/>
+        <location filename="../devicetab.cpp" line="164"/>
         <source>Port</source>
         <translation>端口</translation>
     </message>
@@ -608,47 +608,47 @@
         <translation>设备</translation>
     </message>
     <message>
-        <location filename="../devicetab.cpp" line="128"/>
+        <location filename="../devicetab.cpp" line="129"/>
         <source>NoFlowControl</source>
         <translation>无流控</translation>
     </message>
     <message>
-        <location filename="../devicetab.cpp" line="129"/>
+        <location filename="../devicetab.cpp" line="130"/>
         <source>HardwareControl</source>
         <translation>软件流控</translation>
     </message>
     <message>
-        <location filename="../devicetab.cpp" line="130"/>
+        <location filename="../devicetab.cpp" line="131"/>
         <source>SoftwareControl</source>
         <translation>硬件流控</translation>
     </message>
     <message>
-        <location filename="../devicetab.cpp" line="135"/>
+        <location filename="../devicetab.cpp" line="136"/>
         <source>NoParity</source>
         <translation>无校验</translation>
     </message>
     <message>
-        <location filename="../devicetab.cpp" line="136"/>
+        <location filename="../devicetab.cpp" line="137"/>
         <source>EvenParity</source>
         <translation>偶校验</translation>
     </message>
     <message>
-        <location filename="../devicetab.cpp" line="137"/>
+        <location filename="../devicetab.cpp" line="138"/>
         <source>OddParity</source>
         <translation>奇校验</translation>
     </message>
     <message>
-        <location filename="../devicetab.cpp" line="138"/>
+        <location filename="../devicetab.cpp" line="139"/>
         <source>SpaceParity</source>
         <translation>固定0</translation>
     </message>
     <message>
-        <location filename="../devicetab.cpp" line="139"/>
+        <location filename="../devicetab.cpp" line="140"/>
         <source>MarkParity</source>
         <translation>固定1</translation>
     </message>
     <message>
-        <location filename="../devicetab.cpp" line="276"/>
+        <location filename="../devicetab.cpp" line="279"/>
         <source>Discovered</source>
         <translation>已发现</translation>
     </message>
@@ -661,474 +661,159 @@
         <translation></translation>
     </message>
     <message>
-        <source>Ports</source>
-        <translation type="vanished">端口</translation>
-    </message>
-    <message>
-        <source>Refresh</source>
-        <translation type="vanished">刷新</translation>
-    </message>
-    <message>
-        <source>PortName</source>
-        <translation type="vanished">端口名</translation>
-    </message>
-    <message>
-        <source>Description</source>
-        <translation type="vanished">描述</translation>
-    </message>
-    <message>
-        <source>Manufacturer</source>
-        <translation type="vanished">生产商</translation>
-    </message>
-    <message>
-        <source>SerialNumber</source>
-        <translation type="vanished">序列号</translation>
-    </message>
-    <message>
-        <source>IsNull</source>
-        <translation type="vanished">是否为空</translation>
-    </message>
-    <message>
-        <source>SystemLocation</source>
-        <translation type="vanished">系统路径</translation>
-    </message>
-    <message>
-        <source>VendorID</source>
-        <translation type="vanished">厂商ID</translation>
-    </message>
-    <message>
-        <source>ProductID</source>
-        <translation type="vanished">产品ID</translation>
-    </message>
-    <message>
-        <source>BaudRates</source>
-        <translation type="vanished">波特率</translation>
-    </message>
-    <message>
-        <source>Connect</source>
-        <translation type="vanished">连接</translation>
-    </message>
-    <message>
-        <source>Open</source>
-        <translation type="vanished">打开</translation>
-    </message>
-    <message>
-        <source>Port:</source>
-        <translation type="vanished">端口：</translation>
-    </message>
-    <message>
-        <source>DataBits:</source>
-        <translation type="vanished">数据位：</translation>
-    </message>
-    <message>
-        <source>StopBits:</source>
-        <translation type="vanished">停止位：</translation>
-    </message>
-    <message>
-        <source>BaudRate:</source>
-        <translation type="vanished">波特率：</translation>
-    </message>
-    <message>
-        <source>Parity:</source>
-        <translation type="vanished">校验：</translation>
-    </message>
-    <message>
-        <source>FlowControl:</source>
-        <translation type="vanished">流控：</translation>
-    </message>
-    <message>
-        <source>Advanced</source>
-        <translation type="vanished">高级</translation>
-    </message>
-    <message>
-        <source>Close</source>
-        <translation type="vanished">关闭</translation>
-    </message>
-    <message>
-        <location filename="../mainwindow.cpp" line="71"/>
+        <location filename="../mainwindow.cpp" line="84"/>
         <source>Data</source>
         <translation>数据</translation>
     </message>
     <message>
-        <source>Received:</source>
-        <translation type="vanished">已接收：</translation>
-    </message>
-    <message>
-        <source>Hex</source>
-        <translation type="vanished">16进制</translation>
-    </message>
-    <message>
-        <source>Show latest</source>
-        <translation type="vanished">显示最新数据</translation>
-    </message>
-    <message>
-        <source>Realtime</source>
-        <translation type="vanished">实时</translation>
-    </message>
-    <message>
-        <source>Update</source>
-        <translation type="vanished">更新</translation>
-    </message>
-    <message>
-        <source>Export</source>
-        <translation type="vanished">导出</translation>
-    </message>
-    <message>
-        <source>Copy All</source>
-        <translation type="vanished">全部复制</translation>
-    </message>
-    <message>
-        <source>Clear</source>
-        <translation type="vanished">清空</translation>
-    </message>
-    <message>
-        <source>Sended:</source>
-        <translation type="vanished">已发送：</translation>
-    </message>
-    <message>
-        <source>Send</source>
-        <translation type="vanished">发送</translation>
-    </message>
-    <message>
-        <source>Suffix</source>
-        <translation type="vanished">后缀</translation>
-    </message>
-    <message>
-        <source>String</source>
-        <translation type="vanished">字符串</translation>
-    </message>
-    <message>
-        <source>Repeat</source>
-        <translation type="vanished">重复</translation>
-    </message>
-    <message>
-        <source>Flow Control</source>
-        <translation type="vanished">流控</translation>
-    </message>
-    <message>
-        <source>Encoding</source>
-        <translation type="vanished">编码</translation>
-    </message>
-    <message>
-        <source>Set</source>
-        <translation type="vanished">设置</translation>
-    </message>
-    <message>
-        <location filename="../mainwindow.cpp" line="74"/>
+        <location filename="../mainwindow.cpp" line="87"/>
         <source>Plot</source>
         <translation>绘图</translation>
     </message>
     <message>
-        <source>Enable</source>
-        <translation type="vanished">启动</translation>
-    </message>
-    <message>
-        <source>X Axis:</source>
-        <translation type="vanished">X轴：</translation>
-    </message>
-    <message>
-        <source>Count</source>
-        <translation type="vanished">计数</translation>
-    </message>
-    <message>
-        <source>First data</source>
-        <translation type="vanished">第一项数据</translation>
-    </message>
-    <message>
-        <source>Time</source>
-        <translation type="vanished">时间</translation>
-    </message>
-    <message>
-        <source>Data Num:</source>
-        <translation type="vanished">数据项数：</translation>
-    </message>
-    <message>
-        <source>Legend</source>
-        <translation type="vanished">图例</translation>
-    </message>
-    <message>
-        <source>Fit X</source>
-        <translation type="vanished">适应X轴</translation>
-    </message>
-    <message>
-        <source>Fit Y</source>
-        <translation type="vanished">适应Y轴</translation>
-    </message>
-    <message>
-        <source>Tracer</source>
-        <translation type="vanished">光标</translation>
-    </message>
-    <message>
-        <source>Frame spliter:</source>
-        <translation type="vanished">帧分隔符：</translation>
-    </message>
-    <message>
-        <source>Data Spliter:</source>
-        <translation type="vanished">数据项分隔符：</translation>
-    </message>
-    <message>
-        <source>Clear:</source>
-        <translation type="vanished">清空符：</translation>
-    </message>
-    <message>
-        <source>None</source>
-        <translation type="vanished">无</translation>
-    </message>
-    <message>
-        <source>Scatter</source>
-        <translation type="vanished">显示点</translation>
-    </message>
-    <message>
-        <location filename="../mainwindow.cpp" line="78"/>
+        <location filename="../mainwindow.cpp" line="91"/>
         <source>Control</source>
         <translation>控制</translation>
     </message>
     <message>
-        <source>Add Command</source>
-        <translation type="vanished">添加命令</translation>
-    </message>
-    <message>
-        <source>Add Slider</source>
-        <translation type="vanished">添加滑块</translation>
-    </message>
-    <message>
-        <source>Add CheckBox</source>
-        <translation type="vanished">添加开关</translation>
-    </message>
-    <message>
-        <source>Add Spinbox</source>
-        <translation type="vanished">添加数值框</translation>
-    </message>
-    <message>
-        <source>Import</source>
-        <translation type="vanished">导入</translation>
-    </message>
-    <message>
-        <location filename="../mainwindow.cpp" line="36"/>
+        <location filename="../mainwindow.cpp" line="46"/>
         <source>On Top</source>
         <translation>窗口置顶</translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="43"/>
+        <location filename="../mainwindow.cpp" line="56"/>
         <source>Dock all windows</source>
         <translation>吸附所有悬浮窗口</translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="94"/>
+        <location filename="../mainwindow.cpp" line="107"/>
         <source>Ver: </source>
         <translation>版本： </translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="95"/>
+        <location filename="../mainwindow.cpp" line="108"/>
         <source>Check Update</source>
         <translation>检查更新</translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="141"/>
+        <location filename="../mainwindow.cpp" line="154"/>
         <source>Plz connect to a port first.</source>
         <translation>请先打开串口。</translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="158"/>
-        <location filename="../mainwindow.cpp" line="219"/>
+        <location filename="../mainwindow.cpp" line="171"/>
+        <location filename="../mainwindow.cpp" line="233"/>
         <source>Cannot open the serial port.</source>
         <translation>无法打开串口。</translation>
     </message>
     <message>
-        <source>DeviceName</source>
-        <translation type="vanished">设备名</translation>
-    </message>
-    <message>
-        <source>Type</source>
-        <translation type="vanished">类型</translation>
-    </message>
-    <message>
-        <source>MAC Address</source>
-        <translation type="vanished">MAC地址</translation>
-    </message>
-    <message>
-        <source>NoFlowControl</source>
-        <translation type="vanished">无流控</translation>
-    </message>
-    <message>
-        <source>HardwareControl</source>
-        <translation type="vanished">软件流控</translation>
-    </message>
-    <message>
-        <source>SoftwareControl</source>
-        <translation type="vanished">硬件流控</translation>
-    </message>
-    <message>
-        <location filename="../mainwindow.cpp" line="242"/>
+        <location filename="../mainwindow.cpp" line="256"/>
         <source>NoParity</source>
         <translation>无校验</translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="242"/>
+        <location filename="../mainwindow.cpp" line="256"/>
         <source>EvenParity</source>
         <translation>偶校验</translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="242"/>
+        <location filename="../mainwindow.cpp" line="256"/>
         <source>OddParity</source>
         <translation>奇校验</translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="242"/>
+        <location filename="../mainwindow.cpp" line="256"/>
         <source>SpaceParity</source>
         <translation>固定0</translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="242"/>
+        <location filename="../mainwindow.cpp" line="256"/>
         <source>MarkParity</source>
         <translation>固定1</translation>
     </message>
     <message>
-        <source>Searching...</source>
-        <translation type="vanished">搜索中...</translation>
-    </message>
-    <message>
-        <source>Bonded</source>
-        <translation type="vanished">已配对</translation>
-    </message>
-    <message>
-        <location filename="../mainwindow.cpp" line="241"/>
+        <location filename="../mainwindow.cpp" line="255"/>
         <source>OneStop</source>
         <translation>1</translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="241"/>
+        <location filename="../mainwindow.cpp" line="255"/>
         <source>TwoStop</source>
         <translation>2</translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="241"/>
+        <location filename="../mainwindow.cpp" line="255"/>
         <source>OneAndHalfStop</source>
         <translation>1.5</translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="245"/>
-        <location filename="../mainwindow.cpp" line="252"/>
+        <location filename="../mainwindow.cpp" line="259"/>
+        <location filename="../mainwindow.cpp" line="266"/>
         <source>BaudRate</source>
         <translation>波特率</translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="246"/>
-        <location filename="../mainwindow.cpp" line="253"/>
+        <location filename="../mainwindow.cpp" line="260"/>
+        <location filename="../mainwindow.cpp" line="267"/>
         <source>DataBits</source>
         <translation>数据位</translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="247"/>
-        <location filename="../mainwindow.cpp" line="254"/>
+        <location filename="../mainwindow.cpp" line="261"/>
+        <location filename="../mainwindow.cpp" line="268"/>
         <source>StopBits</source>
         <translation>停止位</translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="248"/>
-        <location filename="../mainwindow.cpp" line="255"/>
+        <location filename="../mainwindow.cpp" line="262"/>
+        <location filename="../mainwindow.cpp" line="269"/>
         <source>Parity</source>
         <translation>校验</translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="259"/>
-        <location filename="../mainwindow.cpp" line="261"/>
+        <location filename="../mainwindow.cpp" line="273"/>
+        <location filename="../mainwindow.cpp" line="275"/>
         <source>State</source>
         <translation>状态</translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="65"/>
-        <location filename="../mainwindow.cpp" line="262"/>
+        <location filename="../mainwindow.cpp" line="78"/>
+        <location filename="../mainwindow.cpp" line="276"/>
         <source>Port</source>
         <translation>端口</translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="263"/>
-        <location filename="../mainwindow.cpp" line="303"/>
+        <location filename="../mainwindow.cpp" line="18"/>
+        <source>Bluetooth is invalid!</source>
+        <translation>蓝牙无法使用！</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="21"/>
+        <source>Please enable Bluetooth!</source>
+        <translation>请先打开蓝牙！</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="277"/>
+        <location filename="../mainwindow.cpp" line="324"/>
         <source>Rx</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="264"/>
+        <location filename="../mainwindow.cpp" line="278"/>
         <source>Tx</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="312"/>
+        <location filename="../mainwindow.cpp" line="18"/>
+        <location filename="../mainwindow.cpp" line="21"/>
+        <location filename="../mainwindow.cpp" line="333"/>
         <source>Error</source>
         <translation>错误</translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="312"/>
+        <location filename="../mainwindow.cpp" line="333"/>
         <source>No port is opened.</source>
         <translation>未打开串口。</translation>
-    </message>
-    <message>
-        <source>Export received data</source>
-        <translation type="vanished">导出已接收数据</translation>
-    </message>
-    <message>
-        <source>Info</source>
-        <translation type="vanished">信息</translation>
-    </message>
-    <message>
-        <source>Successed!</source>
-        <translation type="vanished">成功！</translation>
-    </message>
-    <message>
-        <source>Failed!</source>
-        <translation type="vanished">失败！</translation>
-    </message>
-    <message>
-        <source>Export sended data</source>
-        <translation type="vanished">导出已发送数据</translation>
-    </message>
-    <message>
-        <source>Legend:</source>
-        <translation type="vanished">图例：</translation>
-    </message>
-    <message>
-        <source>New graph name:</source>
-        <translation type="vanished">绘图名称：</translation>
-    </message>
-    <message>
-        <source>Discovered</source>
-        <translation type="vanished">已发现</translation>
-    </message>
-    <message>
-        <source>Paste the exported data in the box.</source>
-        <oldsource># Paste the exported data in the box.</oldsource>
-        <translation type="vanished">将导出的数据粘贴到此框中。</translation>
-    </message>
-    <message>
-        <source>Done</source>
-        <translation type="vanished">完成</translation>
-    </message>
-    <message>
-        <source>Import Control Panel</source>
-        <translation type="vanished">导入控制面板</translation>
-    </message>
-    <message>
-        <source>Please add item first</source>
-        <translation type="vanished">请先添加控制项</translation>
-    </message>
-    <message>
-        <source>Copy all text in this box and save it to somewhere.</source>
-        <oldsource># Copy all text in this box and save it to somewhere.</oldsource>
-        <translation type="vanished">将此文本框中的内容复制并保存。</translation>
-    </message>
-    <message>
-        <source>To import, click the Import button, then paste the text back.</source>
-        <oldsource># To import, click the Import button, then paste the text back.</oldsource>
-        <translation type="vanished">需要导入时，单击“导入”按钮，将复制的内容粘贴回来即可。</translation>
-    </message>
-    <message>
-        <source>Export Control Panel</source>
-        <translation type="vanished">导出控制面板</translation>
-    </message>
-    <message>
-        <source>is not a valid encoding.</source>
-        <translation type="vanished">不是有效编码名。</translation>
     </message>
 </context>
 <context>
@@ -1175,7 +860,7 @@
     </message>
     <message>
         <location filename="../ui/plottab.ui" line="150"/>
-        <location filename="../plottab.cpp" line="94"/>
+        <location filename="../plottab.cpp" line="565"/>
         <source>Legend</source>
         <translation>图例</translation>
     </message>
@@ -1251,17 +936,37 @@
         <translation>显示点</translation>
     </message>
     <message>
-        <source>Legend:</source>
-        <translation type="obsolete">图例：</translation>
-    </message>
-    <message>
-        <source>New graph name:</source>
-        <translation type="obsolete">绘图名称：</translation>
-    </message>
-    <message>
-        <location filename="../plottab.cpp" line="94"/>
+        <location filename="../plottab.cpp" line="565"/>
         <source>New graph name</source>
         <translation>绘图名称</translation>
+    </message>
+</context>
+<context>
+    <name>SerialPinout</name>
+    <message>
+        <location filename="../ui/serialpinout.ui" line="45"/>
+        <source>CTS</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../ui/serialpinout.ui" line="62"/>
+        <source>DSR</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../ui/serialpinout.ui" line="79"/>
+        <source>DCD</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../ui/serialpinout.ui" line="96"/>
+        <source>RI</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../serialpinout.cpp" line="64"/>
+        <source>Pinouts</source>
+        <translation>流控信号</translation>
     </message>
 </context>
 </TS>