+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 
-[中文](doc/CHANGELOG/CHANGELOG_zh_CN.md) 
+[中文](doc/CHANGELOG/CHANGELOG_zh_CN.md)  
+
+## V0.1.2
++ Easier to use  
++ Use Json to storage control items in CtrlTab
++ Save splitter position in DataTab
++ Support serialport pinouts
++ Fix some bugs  
+
 ## V0.1.1
 + Support non-ascii characters in path
 + Support encodings